feat: revise metrics processing
