+++ packages/superset-ui-chart/src/index.ts
@@ -21,10 +21,8 @@ export { default as DatasourceKey } from './query/DatasourceKey';
 
 export { default as ChartDataProvider } from './components/ChartDataProvider';
 
-export * from './types/formData/Annotation';
-export * from './types/Datasource';
 export * from './types/ChartFormData';
+export * from './types/Datasource';
 export * from './types/Query';
+export * from './types/formData/Annotation';
 export * from './types/formData/Metric';
-export * from './types/Operator';
-export * from './types/Time';
