+++ packages/superset-ui-chart/src/query/convertFilter.ts
@@ -1,0 +1,34 @@
+import {
+  SimpleAdhocFilter,
+  isUnaryAdhocFilter,
+  isBinaryAdhocFilter,
+} from '../types/formData/Filter';
+import { QueryObjectFilterClause } from '../types/Query';
+
+export default function convertFilter(filter: SimpleAdhocFilter): QueryObjectFilterClause {
+  const { subject } = filter;
+  if (isUnaryAdhocFilter(filter)) {
+    const { operator } = filter;
+
+    return {
+      col: subject,
+      op: operator,
+    };
+  } else if (isBinaryAdhocFilter(filter)) {
+    const { operator } = filter;
+
+    return {
+      col: subject,
+      op: operator,
+      val: filter.comparator,
+    };
+  }
+
+  const { operator } = filter;
+
+  return {
+    col: subject,
+    op: operator,
+    val: filter.comparator,
+  };
+}
